fix(ml-billing): order_id must be BigInteger, or the first real insert overflows
`MlBillingChargeOrder.order_id` was `Integer`. A Mercado Libre order id is
around 2000018265495500; a Postgres INTEGER stops at 2147483647. Every
`order_id` in `ml_orders_ops.py` is already `BigInteger` — this was the one
that was not.

Nothing would have caught it before production. SQLite does not distinguish
integer widths, so the whole model suite passed and the migration applied
cleanly; the overflow only lands on the first INSERT against Postgres, once
the writer exists two cortes from now.

Verified against the real database: the column is `bigint` and an actual
ML order id inserts and reads back intact.

The regression test asserts the TYPE rather than the behaviour, and says so
in its docstring — an INSERT-based test would pass under SQLite whatever
the width is, so the type declaration is the only guard available here.

+++ backend/alembic/versions/20260904_ml_billing_schema.py
@@ -52,7 +52,7 @@ def upgrade() -> None:
         "ml_billing_charge_orders",
         sa.Column("id", sa.Integer(), nullable=False),
         sa.Column("detail_id", sa.String(length=60), nullable=False),
-        sa.Column("order_id", sa.Integer(), nullable=False),
+        sa.Column("order_id", sa.BigInteger(), nullable=False),
         sa.ForeignKeyConstraint(["detail_id"], ["ml_billing_charges.detail_id"]),
         sa.PrimaryKeyConstraint("id"),
         sa.UniqueConstraint("detail_id", "order_id", name="uq_ml_billing_charge_orders_detail_order"),
